Move language types to js.go and kinds to kinds.go
