+++ gazelle/js/js.go
@@ -1,0 +1,37 @@
+// Copyright 2019 The Bazel Authors. All rights reserved.
+// Modifications copyright (C) 2021 BenchSci Analytics Inc.
+// Modifications copyright (C) 2018 Ecosia GmbH
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+
+// http://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package js
+
+import (
+	"github.com/bazelbuild/bazel-gazelle/language"
+)
+
+const languageName = "js"
+
+// Name returns the name of the language. This should be a prefix of the
+// kinds of rules generated by the language, e.g., "go" for the Go extension
+// since it generates "go_library" rules.
+func (*JS) Name() string {
+	return "js"
+}
+
+type JS struct {
+}
+
+func NewLanguage() language.Language {
+	return &JS{}
+}
