+++ gazelle/js/lang.go
@@ -29,112 +29,12 @@ import (
 	"github.com/bazelbuild/bazel-gazelle/rule"
 )
 
-const languageName = "js"
-
-// Name returns the name of the language. This should be a prefix of the
-// kinds of rules generated by the language, e.g., "go" for the Go extension
-// since it generates "go_library" rules.
-func (*JS) Name() string {
-	return "js"
-}
-
-type JS struct {
-}
-
-func NewLanguage() language.Language {
-	return &JS{}
-}
-
 type imports struct {
 	set map[string]bool
 }
 
 var noImports = imports{
 	set: map[string]bool{},
-}
-
-// Kinds returns a map of maps rule names (kinds) and information on how to
-// match and merge attributes that may be found in rules of those kinds. All
-// kinds of rules generated for this language may be found here.
-func (*JS) Kinds() map[string]rule.KindInfo {
-	return map[string]rule.KindInfo{
-		"js_library": {
-			MatchAny: false,
-			NonEmptyAttrs: map[string]bool{
-				"srcs": true,
-			},
-			MergeableAttrs: map[string]bool{
-				"srcs": true,
-				"tags": true,
-			},
-			ResolveAttrs: map[string]bool{
-				"deps": true,
-				"data": true,
-			},
-		},
-		"ts_project": {
-			MatchAny: false,
-			NonEmptyAttrs: map[string]bool{
-				"srcs": true,
-			},
-			MergeableAttrs: map[string]bool{
-				"srcs": true,
-				"tags": true,
-			},
-			ResolveAttrs: map[string]bool{
-				"deps": true,
-				"data": true,
-			},
-		},
-		"ts_definition": {
-			MatchAny: false,
-			NonEmptyAttrs: map[string]bool{
-				"srcs": true,
-			},
-			MergeableAttrs: map[string]bool{
-				"srcs": true,
-				"tags": true,
-			},
-			ResolveAttrs: map[string]bool{
-				"deps": true,
-				"data": true,
-			},
-		},
-		"jest_test": {
-			MatchAny: false,
-			NonEmptyAttrs: map[string]bool{
-				"srcs": true,
-			},
-			MergeableAttrs: map[string]bool{
-				"srcs": true,
-				"tags": true,
-			},
-			ResolveAttrs: map[string]bool{
-				"deps": true,
-				"data": true,
-			},
-		},
-		"web_asset": {
-			MatchAny: false,
-			NonEmptyAttrs: map[string]bool{
-				"srcs": true,
-			},
-			MergeableAttrs: map[string]bool{
-				"srcs": true,
-				"tags": true,
-			},
-		},
-		"web_assets": {
-			MatchAny: false,
-			NonEmptyAttrs: map[string]bool{
-				"srcs": true,
-			},
-			MergeableAttrs: map[string]bool{
-				"srcs": true,
-				"tags": true,
-			},
-		},
-	}
 }
 
 var managedRules = []string{"js_library", "ts_project", "jest_test", "web_asset", "web_assets", "ts_definition"}
